Remove all node code from browser build
