--- src/util/is-browser.js
@@ -1,3 +1,0 @@
-module.exports = function isBrowser () {
-  return (typeof window !== 'undefined')
-}
