+++ src/util/is-color-supported/browser.js
@@ -1,0 +1,6 @@
+var isWebkit = require('../is-webkit')
+var isFirefox = require('../is-firefox')
+
+module.exports = function isColorSupported () {
+  return (isWebkit() || isFirefox())
+}
