+++ test/browser/index.test.js
@@ -1,8 +1,7 @@
 /* eslint-env jest */
 
 jest.mock('../../src/util/is-webkit', () => () => true)
-jest.mock('../../src/util/is-browser', () => () => true)
-jest.mock('../../src/util/is-color-supported', () => () => true)
+jest.mock('../../src/util/is-color-supported/browser', () => () => true)
 
 describe('browser', () => {
   require('./local-storage-debug-var')
