+++ test/node/index.test.js
@@ -1,7 +1,5 @@
 /* eslint-env jest */
 
-jest.mock('../../src/util/is-browser', () => () => false)
-
 describe('node', () => {
   require('./logging-methods')
   require('./debug-env-var')
